Implement binary protocol layering over HTTP
