+++ package.json
@@ -28,6 +28,7 @@
     "@types/node": "14",
     "get-stdin": "^7.0.0",
     "jest": "^26.6.3",
+    "node-fetch": "2.6.7",
     "nodemon": "^2.0.13",
     "prettier": "^2.3.2",
     "proposal-temporal": "^0.7.0",
