+++ src/codecs/ifaces.ts
@@ -37,7 +37,6 @@ export interface ICodec {
   decode(buf: ReadBuffer): any;
 
   getSubcodecs(): ICodec[];
-  getSubcodecsNames(): string[];
   getKind(): CodecKind;
   getKnownTypeName(): string;
 }
@@ -57,10 +56,6 @@ export abstract class Codec {
 
   getKnownTypeName(): string {
     return "anytype";
-  }
-
-  getSubcodecsNames(): string[] {
-    return [];
   }
 }
